Fixed ReqRelCopier
traversal of relationship types now occur from the requirements themselves
and are no longer dependent on using SCOPE to find owned relationships we
want to operate on.  SCOPE is now provided as the owner of newly created
relationships.  If it is desired to have new relationship owners in specific
packages, then separate runs per relationship type can be conducted (or even
separate runs per owner), manually.